\name{packizaskun-package}
\alias{packizaskun-package}
\alias{packizaskun}
\docType{package}
\title{
\packageTitle{packizaskun}
}
\description{
\packageDescription{packizaskun}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{packizaskun}
\packageIndices{packizaskun}
~~ Contiene tres funciones con operaciones sencillas sumar.dos.numeros, sumar.tres.numeros, restar.dos.numeros ~~
}
\author{
\packageAuthor{packizaskun}

Maintainer: \packageMaintainer{packizaskun}
}
\references{
~~ No hay referencias ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
